--- src/application/hooks/api/account/queryKey.ts
@@ -1,3 +1,0 @@
-export const QUERY_KEYS = {
-  getMyAccount: ["getMyAccount"],
-} as const;
