--- src/application/hooks/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./account";
